export_savedmodel {tensorflow}	R Documentation

Export a Saved Model

Description

Serialize a model to disk. See implementations in the keras and tfestimators packages.

Usage

export_savedmodel(object, export_dir_base, ...)

Arguments

object
An R object.

export_dir_base
A string containing a directory in which to export the SavedModel.

...
Optional arguments passed on to implementing methods.

Value

The path to the exported directory, as a string.

Implementations

keras

tfestimators

[Package tensorflow version 2.5.0 Index]