Install TensorFlow and it's dependencies

Install TensorFlow and it's dependencies

install_tensorflow(method = c("auto", "virtualenv", "conda", "system"),
  version = "latest", gpu = FALSE, package_url = NULL, conda = "auto")

Arguments

method
Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows (as this isn't supported by TensorFlow). Note also that since this command runs without privillege the "system" method is available only on Windows.

version
TensorFlow version to install (must be either "latest" or a full major.minor.patch specification, e.g. "1.1.0").

gpu
Install the GPU version of TensorFlow

package_url
URL of the TensorFlow package to install (if not specified this is determined automatically). Note that if this parameter is provied then the version and gpu parameters are ignored.

conda
Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).