parse_arguments {tensorflow}	R Documentation

Parse Command Line Arguments

Description

Parse command line arguments of the form --key=value and --key value. The values are assumed to be valid yaml and will be converted using yaml.load().

Usage

parse_arguments(arguments = NULL)

Arguments

arguments
A vector of command line arguments. When NULL (the default), the command line arguments received by the current R process are used.

[Package tensorflow version 2.5.0 Index]