tf_config {tensorflow}	R Documentation

TensorFlow configuration information

Description

TensorFlow configuration information

Usage

tf_config()

tf_version()

Value

List with information on the current configuration of TensorFlow. You can determine whether TensorFlow was found using the available member (other members vary depending on whether available is TRUE or FALSE)

[Package tensorflow version 2.5.0 Index]