tf_function {tensorflow}	R Documentation

Creates a callable TensorFlow graph from an R function.

Description

tf_function constructs a callable that executes a TensorFlow graph created by tracing the TensorFlow operations in f. This allows the TensorFlow runtime to apply optimizations and exploit parallelism in the computation defined by f.

Usage

tf_function(f, input_signature = NULL, autograph = FALSE, ...)

Arguments

f
the function to be compiled

input_signature
A possibly nested sequence of tf$TensorSpec objects specifying the shapes and dtypes of the tensors that will be supplied to this function. If NULL, a separate function is instantiated for each inferred input signature. If input_signature is specified, every input to f must be a tensor.

autograph
Whether autograph should be applied on f before tracing a graph. This allows for dynamic control flow (if's, loops etc.) in the traced graph. See https://www.tensorflow.org/guide/autograph for more information. Note: We set the default to FALSE until this functionality is available from R.

...
additional arguments passed on to tf.function (vary based on Tensorflow version). See https://www.tensorflow.org/api_docs/python/tf/function for details.

[Package tensorflow version 2.5.0 Index]